% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gauges.R
\name{get_gauges}
\alias{get_gauges}
\title{Get gauge metadata (simplified), locations and latest measurements}
\usage{
get_gauges()
}
\value{
Sf object.
}
\description{
Get gauge metadata (simplified), locations and latest measurements
}
\examples{
\dontrun{
get_gauges()
}
}
