% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Get water level or discharge measurements for selected gauges}
\usage{
get_measurements(x = NULL, discharge = FALSE)
}
\arguments{
\item{x}{Sf object containing gauges to be used for subsequent queries,
as provided by \code{get_gauges()}.}

\item{discharge}{logical. Retrieve water level or discharge data?}
}
\value{
List of xts objects.
}
\description{
Get water level or discharge measurements for selected gauges
}
\examples{
\dontrun{
gauge <- get_gauges() |> dplyr::filter(id == "10103")

get_measurements(gauge)
get_measurements(gauge, discharge = TRUE)

gauges <- get_gauges() |> dplyr::filter(waterbody == "Hammbach")

get_measurements(gauges)
}
}
\seealso{
\code{\link[=get_gauges]{get_gauges()}}
}
