% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get gauge metadata}
\usage{
get_meta(x = NULL)
}
\arguments{
\item{x}{character. Station ID.}
}
\value{
Tibble.
}
\description{
Get gauge metadata
}
\examples{
get_meta("10101")
}
