% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rv_gauges.R
\name{get_rv_gauges}
\alias{get_rv_gauges}
\title{Get RV gauge metadata (simplified), locations and latest measurements}
\usage{
get_rv_gauges()
}
\value{
Sf object.
}
\description{
Get RV gauge metadata (simplified), locations and latest measurements
}
\examples{
\dontrun{
get_rv_gauges()
}
}
