% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Get water level or discharge measurements}
\usage{
get_measurements(x = NULL, discharge = FALSE)
}
\arguments{
\item{x}{character. Station ID.}

\item{discharge}{logical. Retrieve water level or discharge data?}
}
\value{
xts object.
}
\description{
Get water level or discharge measurements
}
\examples{
get_measurements("10104")

get_measurements("10103", discharge = TRUE)
}
