% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get (extended) metadata for selected gauges}
\usage{
get_meta(x = NULL)
}
\arguments{
\item{x}{Sf object containing gauges to be used for subsequent queries,
as provided by \code{get_gauges()}.}
}
\value{
Tibble containing metadata.
}
\description{
Get (extended) metadata for selected gauges
}
\examples{
\dontrun{
gauge <- get_gauges() |> dplyr::filter(id == "10103")
get_meta(gauge)

gauges <- get_gauges() |> dplyr::filter(waterbody == "Hammbach")
get_meta(gauges)
}
}
\seealso{
\code{\link[=get_gauges]{get_gauges()}}
}
